Consolidate vision helpers
